Removed a needless conversion to double

+++ stats/stats.scala
@@ -49,10 +49,8 @@ object Stats {
     lazy val ratioAvg: Double = if (total == 0) 0.0 else coveredAvg / total
     lazy val percentageAvg: Double = 100.0 * ratioAvg
     lazy val stdDev: Double = Math.sqrt(
-      (covered.map{ c => (c - coveredAvg) * (c - coveredAvg) }
-        .sum
-        .toDouble
-      ) / covered.length
+      covered.map{ c => (c - coveredAvg) * (c - coveredAvg) }.sum /
+        covered.length
     )
 
     override def toString: String = Seq(
